+++ src/HowToOptimizeGemm/MMult1.c
@@ -6,11 +6,11 @@
 
 /* Routine for computing C = A * B + C */
 
-void AddDot( int, double *, int, double *, double * );
+void AddDot( int, float *, int, float *, float * );
 
-void MY_MMult( int m, int n, int k, double *a, int lda, 
-                                    double *b, int ldb,
-                                    double *c, int ldc )
+void MY_MMult( int m, int n, int k, float *a, int lda, 
+                                    float *b, int ldb,
+                                    float *c, int ldc )
 {
   int i, j;
 
@@ -29,7 +29,7 @@ void MY_MMult( int m, int n, int k, double *a, int lda,
 
 #define X(i) x[ (i)*incx ]
 
-void AddDot( int k, double *x, int incx,  double *y, double *gamma )
+void AddDot( int k, float *x, int incx,  float *y, float *gamma )
 {
   /* compute gamma := x' * y + gamma with vectors x and y of length n.
